+++ Sources/APNSwiftCore/APNSPriority.swift
