+++ Sources/APNSwiftCore/APNSwift.docc/APNSwift.md
@@ -72,7 +72,7 @@ try await client.sendAlertNotification(
         payload: Payload()
     ),
     deviceToken: "device-token",
-    deadline: .distantFuture,
+    deadline: .nanoseconds(Int64.max),
     logger: myLogger
 )
 ```
