+++ Sources/APNSwiftCore/LiveActivity/APNSClient+LiveActivity.swift
@@ -12,33 +12,33 @@
 //
 //===----------------------------------------------------------------------===//
 
-import NIOCore
 import Logging
 
 extension APNSClient {
-    /// Sends a VoIP notification to APNs.
+    /// Sends a live activity notification.
     ///
     /// - Parameters:
     ///   - notification: The notification to send.
     ///
-    ///   - deviceToken: The hexadecimal bytes that identify the user’s device. Your app receives the bytes for this device token
-    ///    when registering for remote notifications.
+    ///   - deviceToken: The hexadecimal bytes use to send live activity notification. Your app receives the bytes for this activity token
+    ///    from `pushTokenUpdates` async property of a live activity.
     ///
     ///   - deadline: Point in time by which sending the notification to APNs must complete.
     ///
     ///   - logger: The logger to use for sending this notification.
     @discardableResult
     @inlinable
-    public func sendVoIPNotification<Payload: Encodable>(
-        _ notification: APNSVoIPNotification<Payload>,
+    public func sendLiveActivityNotification<ContentState: Encodable>(
+        _ notification: APNSLiveActivityNotification<ContentState>,
         deviceToken: String,
-        deadline: NIODeadline,
+        deadline: Duration,
         logger: Logger = _noOpLogger
     ) async throws -> APNSResponse {
-        try await self.send(
-            payload: notification.payload,
+        return try await self.send(
+            payload: notification,
             deviceToken: deviceToken,
-            pushType: .voip,
+            pushType: .liveactivity,
+            apnsID: notification.apnsID,
             expiration: notification.expiration,
             priority: notification.priority,
             topic: notification.topic,
@@ -47,3 +47,4 @@ extension APNSClient {
         )
     }
 }
+
