+++ Sources/APNSwiftCore/LiveActivity/APNSLiveActivityNotificationEvent.swift
