+++ Sources/APNSwiftCore/Location/APNSLocationNotification.swift
