+++ Sources/APNSwiftCore/VoIP/APNSClient+VoIP.swift
@@ -12,34 +12,32 @@
 //
 //===----------------------------------------------------------------------===//
 
-import NIOCore
 import Logging
 
 extension APNSClient {
-    /// Sends a live activity notification.
+    /// Sends a VoIP notification to APNs.
     ///
     /// - Parameters:
     ///   - notification: The notification to send.
     ///
-    ///   - deviceToken: The hexadecimal bytes use to send live activity notification. Your app receives the bytes for this activity token
-    ///    from `pushTokenUpdates` async property of a live activity.
+    ///   - deviceToken: The hexadecimal bytes that identify the user’s device. Your app receives the bytes for this device token
+    ///    when registering for remote notifications.
     ///
     ///   - deadline: Point in time by which sending the notification to APNs must complete.
     ///
     ///   - logger: The logger to use for sending this notification.
     @discardableResult
     @inlinable
-    public func sendLiveActivityNotification<ContentState: Encodable>(
-        _ notification: APNSLiveActivityNotification<ContentState>,
+    public func sendVoIPNotification<Payload: Encodable>(
+        _ notification: APNSVoIPNotification<Payload>,
         deviceToken: String,
-        deadline: NIODeadline,
+        deadline: Duration,
         logger: Logger = _noOpLogger
     ) async throws -> APNSResponse {
-        return try await self.send(
-            payload: notification,
+        try await self.send(
+            payload: notification.payload,
             deviceToken: deviceToken,
-            pushType: .liveactivity,
-            apnsID: notification.apnsID,
+            pushType: .voip,
             expiration: notification.expiration,
             priority: notification.priority,
             topic: notification.topic,
@@ -48,4 +46,3 @@ extension APNSClient {
         )
     }
 }
-
